refactor(server): split stream-management websocket tests
Move the XEP-0198 stream-management and detached-session WebSocket test coverage out of the oversized parent route test module into a dedicated child module. The split keeps shared fixtures in the parent and only widens imports where the Rust test-module boundary needs them.

Constraint: Follow-up slice for #350; no production WebSocket, XMPP, or XEP wire behavior changes.

Rejected: Move unrelated XEP-0237 and RFC 6121 feature-advertisement tests | those are not stream-management coverage and should stay with the remaining feature-advertisement group.

Confidence: high

Scope-risk: narrow

Directive: Keep XEP-0198 WebSocket regression coverage in tests/stream_management.rs when adding new SM enable, resume, replay, detach, or janitor cases.

Tested: cargo fmt

Tested: cargo test -p waddle-server stream_management::

Tested: cargo test -p waddle-server server::routes::websocket

Tested: cargo clippy -p waddle-server --tests -- -D warnings

Tested: cargo test -p waddle-server

Tested: git diff --check
