+++ server/crates/waddle-server/src/server/routes/websocket/tests/stream_management.rs
@@ -1,0 +1,2270 @@
+use super::super::{
+    cleanup::{cleanup_connection_shutdown, cleanup_muc_presence_for_jid},
+    frame::handle_xmpp_frame,
+    handlers::{self, presence::handle_muc_join},
+    interpret_loop::build_interpret_deps,
+    replay::drive_interpret_loop,
+    state::WsConnState,
+    stream_management::is_countable_stanza,
+    transport_xml::{build_stream_features_xml, element_to_xml, sasl_success_xml, stanza_to_xml},
+};
+use super::{
+    create_test_server_owner_session, create_test_session, create_test_websocket_state,
+    message_frame_xml_with_id, register_test_native_user, scram_client_final_from_challenge,
+    snapshot_room,
+};
+use crate::auth::Session;
+use base64::{engine::general_purpose::STANDARD as BASE64_STANDARD, Engine as _};
+use jid::{BareJid, FullJid};
+use std::str::FromStr;
+use tokio::sync::mpsc;
+use waddle_xmpp::{
+    protocol::{Blocklist, ConnectionPhase, InboundEvent},
+    registry::{DeliveryKind, OutboundStanza},
+    stream_management::{SmSessionRegistry, SM_NS},
+    Stanza,
+};
+use xmpp_parsers::minidom::Element;
+
+fn resume_frame_xml(stream_id: &str, handled_count: u32) -> String {
+    element_to_xml(
+        Element::builder("resume", SM_NS)
+            .attr("previd", stream_id)
+            .attr("h", handled_count.to_string())
+            .build(),
+    )
+}
+
+// ---- XEP-0198 stream management --------------------------------
+
+#[tokio::test]
+async fn sm_features_advertise_sm_namespace() {
+    // Stream features after successful auth must include <sm/>.
+    let features = build_stream_features_xml(true);
+    let el = Element::from_str(&features).expect("features xml");
+    assert!(
+        el.children()
+            .any(|child| child.name() == "sm" && child.ns() == SM_NS),
+        "post-auth features must advertise urn:xmpp:sm:3"
+    );
+}
+
+#[test]
+fn is_countable_stanza_matches_element_name_not_prefix() {
+    // Real stanzas that must count toward SM handled/sent counters.
+    assert!(is_countable_stanza(
+        "<iq xmlns='jabber:client' type='get' id='1'/>"
+    ));
+    assert!(is_countable_stanza("<message xmlns='jabber:client'/>"));
+    assert!(is_countable_stanza("<presence xmlns='jabber:client'/>"));
+    // Leading whitespace is tolerated (matches the pre-existing
+    // trim behaviour — frames are always serialized with a
+    // namespace by minidom, so callers never produce bare `<iq/>`).
+    assert!(is_countable_stanza("  <iq xmlns='jabber:client' id='1'/>"));
+
+    // SM control nonzas and stream-level frames must NOT count.
+    assert!(!is_countable_stanza("<r xmlns='urn:xmpp:sm:3'/>"));
+    assert!(!is_countable_stanza("<a xmlns='urn:xmpp:sm:3' h='1'/>"));
+    assert!(!is_countable_stanza(
+        "<enable xmlns='urn:xmpp:sm:3' resume='1'/>"
+    ));
+    assert!(!is_countable_stanza(
+        "<resumed xmlns='urn:xmpp:sm:3' previd='x' h='0'/>"
+    ));
+
+    // Substring prefix collisions that the old `starts_with`
+    // implementation would have accepted. These are all non-standard
+    // today but the element-name match is how we stay safe if any
+    // future XEP introduces similarly-named nonzas.
+    assert!(!is_countable_stanza("<messages xmlns='urn:example'/>"));
+    assert!(!is_countable_stanza("<presences xmlns='urn:example'/>"));
+    assert!(!is_countable_stanza("<iqsomething/>"));
+
+    // Malformed XML just doesn't count — no panic, no false positive.
+    assert!(!is_countable_stanza("not-xml-at-all"));
+    assert!(!is_countable_stanza(""));
+}
+
+#[tokio::test]
+async fn handle_xmpp_frame_drops_oversized_sm_nonza_before_parse() {
+    let state = create_test_websocket_state().await;
+    let mut conn = WsConnState::new();
+    let huge = element_to_xml(
+        Element::builder("r", SM_NS)
+            .attr(
+                "note",
+                "a".repeat(waddle_xmpp::protocol::frame::MAX_FRAME_SIZE),
+            )
+            .build(),
+    );
+
+    let responses = handle_xmpp_frame(&huge, "example.com", state.as_ref(), &mut conn).await;
+
+    assert!(responses.is_empty());
+}
+
+#[tokio::test]
+async fn sm_enable_requires_resource_binding() {
+    let state = create_test_websocket_state().await;
+    let mut conn = WsConnState::new();
+    // Without resource_bound, enable must fail.
+    let frame = "<enable xmlns='urn:xmpp:sm:3' resume='true'/>";
+    let responses = handle_xmpp_frame(frame, "example.com", state.as_ref(), &mut conn).await;
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "failed");
+    assert!(!conn.sm_state.enabled);
+}
+
+#[tokio::test]
+async fn sm_resume_requires_authentication() {
+    let state = create_test_websocket_state().await;
+    let mut conn = WsConnState::new();
+
+    let resume_frame = resume_frame_xml("stream-xyz", 0);
+    let responses =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "failed");
+    assert!(el
+        .get_child("unexpected-request", "urn:ietf:params:xml:ns:xmpp-stanzas")
+        .is_some());
+    assert!(matches!(conn.phase, ConnectionPhase::Unauthenticated));
+}
+
+#[tokio::test]
+async fn sm_resume_is_rejected_during_scram_and_scram_can_still_complete() {
+    let state = create_test_websocket_state().await;
+    let domain = state.deps.auth_state.xmpp_domain.clone();
+    let password = "correct horse battery staple";
+    let client_nonce = "fyko+d2lbbFgONRv9qkxdawL";
+    register_test_native_user(state.as_ref(), "alice", password).await;
+
+    let auth_frame = element_to_xml(
+        Element::builder("auth", waddle_xmpp::ns::SASL)
+            .attr("mechanism", "SCRAM-SHA-256")
+            .append(BASE64_STANDARD.encode(format!("n,,n=alice,r={client_nonce}")))
+            .build(),
+    );
+    let mut conn = WsConnState::new();
+
+    let auth_responses = handle_xmpp_frame(&auth_frame, &domain, state.as_ref(), &mut conn).await;
+    let challenge = Element::from_str(&auth_responses[0]).expect("challenge xml");
+    let challenge_b64 = challenge.text();
+    assert_eq!(conn.phase.scram_pending_username(), Some("alice"));
+
+    let resume_frame = resume_frame_xml("stream-xyz", 0);
+    let resume_responses =
+        handle_xmpp_frame(&resume_frame, &domain, state.as_ref(), &mut conn).await;
+    assert_eq!(resume_responses.len(), 1);
+    let failed = Element::from_str(&resume_responses[0]).expect("failed xml");
+    assert_eq!(failed.name(), "failed");
+    assert!(failed
+        .get_child("unexpected-request", "urn:ietf:params:xml:ns:xmpp-stanzas")
+        .is_some());
+    assert_eq!(conn.phase.scram_pending_username(), Some("alice"));
+
+    let response_frame = element_to_xml(
+        Element::builder("response", waddle_xmpp::ns::SASL)
+            .append(BASE64_STANDARD.encode(scram_client_final_from_challenge(
+                "alice",
+                password,
+                client_nonce,
+                &challenge_b64,
+            )))
+            .build(),
+    );
+    let response_responses =
+        handle_xmpp_frame(&response_frame, &domain, state.as_ref(), &mut conn).await;
+
+    assert_eq!(response_responses.len(), 1);
+    let success = Element::from_str(&response_responses[0]).expect("success xml");
+    assert_eq!(success.name(), "success");
+    assert!(matches!(conn.phase, ConnectionPhase::Authenticated { .. }));
+    assert!(conn.phase.is_authenticated());
+}
+
+#[tokio::test]
+async fn sm_resume_is_allowed_after_auth_before_bind() {
+    let state = create_test_websocket_state().await;
+    let session = create_test_session(state.as_ref(), "alice").await;
+    let payload = BASE64_STANDARD.encode(format!("n,,\x01auth=Bearer {}\x01\x01", session.id));
+    let auth_frame = element_to_xml(
+        Element::builder("auth", waddle_xmpp::ns::SASL)
+            .attr("mechanism", "OAUTHBEARER")
+            .append(payload)
+            .build(),
+    );
+    let mut conn = WsConnState::new();
+
+    let auth_responses =
+        handle_xmpp_frame(&auth_frame, "example.com", state.as_ref(), &mut conn).await;
+    assert_eq!(auth_responses, vec![sasl_success_xml()]);
+
+    let resume_frame = resume_frame_xml("stream-xyz", 0);
+    let responses =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "failed");
+    assert!(el
+        .get_child("item-not-found", "urn:ietf:params:xml:ns:xmpp-stanzas")
+        .is_some());
+    assert!(matches!(conn.phase, ConnectionPhase::Authenticated { .. }));
+    assert!(!conn.phase.is_resumed());
+}
+
+#[tokio::test]
+async fn sm_resume_rejects_when_replay_window_has_gap() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+
+    let state = create_test_websocket_state().await;
+    let domain = state.deps.auth_state.xmpp_domain.clone();
+    let session = create_test_session(state.as_ref(), "alice").await;
+    let payload = BASE64_STANDARD.encode(format!("n,,\x01auth=Bearer {}\x01\x01", session.id));
+    let auth_frame = element_to_xml(
+        Element::builder("auth", waddle_xmpp::ns::SASL)
+            .attr("mechanism", "OAUTHBEARER")
+            .append(payload)
+            .build(),
+    );
+    let mut conn = WsConnState::new();
+    let auth_responses = handle_xmpp_frame(&auth_frame, &domain, state.as_ref(), &mut conn).await;
+    assert_eq!(auth_responses, vec![sasl_success_xml()]);
+
+    let mut detached = DetachedSession {
+        stream_id: "stream-gap".to_string(),
+        user_id: format!("alice@{domain}"),
+        jid: format!("alice@{domain}/web").parse().expect("jid"),
+        inbound_count: 5,
+        outbound_count: 0,
+        last_acked: 0,
+        replay_gap_through: None,
+        unacked_stanzas: Vec::new(),
+        max_resume_time: Some(300),
+        detached_at: std::time::Instant::now(),
+        carbons_enabled: false,
+        roster_interested: false,
+        presence_available: false,
+        presence_show: None,
+        presence_status: None,
+        presence_priority: 0,
+    };
+    for sequence in 1..=(waddle_xmpp::stream_management::DEFAULT_MAX_UNACKED_QUEUE_SIZE as u32 + 1)
+    {
+        detached.record_detached_outbound_at(
+            sequence,
+            message_frame_xml_with_id(format!("m{sequence}")),
+            chrono::Utc::now(),
+        );
+    }
+    assert_eq!(detached.replay_gap_through, Some(1));
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(detached.clone())
+        .await
+        .expect("store");
+
+    let resume_frame = resume_frame_xml("stream-gap", 0);
+    let responses = handle_xmpp_frame(&resume_frame, &domain, state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "failed");
+    assert_eq!(el.attr("h"), Some("5"));
+    assert!(el
+        .get_child("resource-constraint", "urn:ietf:params:xml:ns:xmpp-stanzas")
+        .is_some());
+    assert!(matches!(conn.phase, ConnectionPhase::Authenticated { .. }));
+    assert!(!conn.phase.is_resumed());
+
+    let stored = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session("stream-gap")
+        .await
+        .expect("take")
+        .expect("detached session should remain for expiry/fallback handling");
+    assert_eq!(stored.jid, detached.jid);
+}
+
+#[tokio::test]
+async fn sm_resume_rejects_authenticated_identity_mismatch_and_preserves_session() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+
+    let state = create_test_websocket_state().await;
+    let domain = state.deps.auth_state.xmpp_domain.clone();
+    let session = create_test_session(state.as_ref(), "bob").await;
+    let payload = BASE64_STANDARD.encode(format!("n,,\x01auth=Bearer {}\x01\x01", session.id));
+    let auth_frame = element_to_xml(
+        Element::builder("auth", waddle_xmpp::ns::SASL)
+            .attr("mechanism", "OAUTHBEARER")
+            .append(payload)
+            .build(),
+    );
+    let mut conn = WsConnState::new();
+    let auth_responses = handle_xmpp_frame(&auth_frame, &domain, state.as_ref(), &mut conn).await;
+    assert_eq!(auth_responses, vec![sasl_success_xml()]);
+    assert!(matches!(conn.phase, ConnectionPhase::Authenticated { .. }));
+
+    let detached = DetachedSession {
+        stream_id: "stream-auth-mismatch".to_string(),
+        user_id: format!("alice@{domain}"),
+        jid: format!("alice@{domain}/web").parse().expect("jid"),
+        inbound_count: 0,
+        outbound_count: 0,
+        last_acked: 0,
+        replay_gap_through: None,
+        unacked_stanzas: Vec::new(),
+        max_resume_time: Some(300),
+        detached_at: std::time::Instant::now(),
+        carbons_enabled: false,
+        roster_interested: false,
+        presence_available: false,
+        presence_show: None,
+        presence_status: None,
+        presence_priority: 0,
+    };
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(detached.clone())
+        .await
+        .expect("store");
+
+    let resume_frame = resume_frame_xml("stream-auth-mismatch", 0);
+    let responses = handle_xmpp_frame(&resume_frame, &domain, state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "failed");
+    assert!(el
+        .get_child("not-authorized", "urn:ietf:params:xml:ns:xmpp-stanzas")
+        .is_some());
+    assert!(matches!(conn.phase, ConnectionPhase::Authenticated { .. }));
+    assert_eq!(
+        conn.phase.authenticated_bare_jid().map(ToString::to_string),
+        Some(format!("bob@{domain}"))
+    );
+
+    let stored = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session("stream-auth-mismatch")
+        .await
+        .expect("take")
+        .expect("detached session should remain");
+    assert_eq!(stored.jid, detached.jid);
+}
+
+#[tokio::test]
+async fn sm_resume_matching_authenticated_identity_preserves_current_session_without_sidecar() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+
+    let state = create_test_websocket_state().await;
+    let domain = state.deps.auth_state.xmpp_domain.clone();
+    let session = create_test_session(state.as_ref(), "bob").await;
+    let payload = BASE64_STANDARD.encode(format!("n,,\x01auth=Bearer {}\x01\x01", session.id));
+    let auth_frame = element_to_xml(
+        Element::builder("auth", waddle_xmpp::ns::SASL)
+            .attr("mechanism", "OAUTHBEARER")
+            .append(payload)
+            .build(),
+    );
+    let mut conn = WsConnState::new();
+    let auth_responses = handle_xmpp_frame(&auth_frame, &domain, state.as_ref(), &mut conn).await;
+    assert_eq!(auth_responses, vec![sasl_success_xml()]);
+    assert!(matches!(conn.phase, ConnectionPhase::Authenticated { .. }));
+
+    let detached_jid: FullJid = format!("bob@{domain}/web").parse().expect("jid");
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: "stream-auth-match".to_string(),
+            user_id: format!("bob@{domain}"),
+            jid: detached_jid.clone(),
+            inbound_count: 2,
+            outbound_count: 3,
+            last_acked: 3,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: true,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store");
+
+    let resume_frame = resume_frame_xml("stream-auth-match", 3);
+    let responses = handle_xmpp_frame(&resume_frame, &domain, state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    let resumed = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(resumed.name(), "resumed");
+    assert_eq!(conn.phase.bound_jid(), Some(&detached_jid));
+    assert!(conn.phase.is_ready());
+    assert!(conn.phase.is_resumed());
+    assert!(matches!(
+        &conn.phase,
+        ConnectionPhase::Ready {
+            full_jid,
+            resumed: true,
+            ..
+        } if full_jid == &detached_jid
+    ));
+    assert_eq!(
+        conn.authenticated_session
+            .as_ref()
+            .map(|saved| saved.user_id.as_str()),
+        Some(session.user_id.as_str())
+    );
+}
+
+#[tokio::test]
+async fn sm_resume_matching_authenticated_identity_prefers_detached_sidecar_session() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+
+    let state = create_test_websocket_state().await;
+    let domain = state.deps.auth_state.xmpp_domain.clone();
+    let fresh_session = create_test_session(state.as_ref(), "bob").await;
+    let payload =
+        BASE64_STANDARD.encode(format!("n,,\x01auth=Bearer {}\x01\x01", fresh_session.id));
+    let auth_frame = element_to_xml(
+        Element::builder("auth", waddle_xmpp::ns::SASL)
+            .attr("mechanism", "OAUTHBEARER")
+            .append(payload)
+            .build(),
+    );
+    let mut conn = WsConnState::new();
+    let auth_responses = handle_xmpp_frame(&auth_frame, &domain, state.as_ref(), &mut conn).await;
+    assert_eq!(auth_responses, vec![sasl_success_xml()]);
+
+    let stream_id = "stream-auth-match-with-sidecar";
+    let detached_jid: FullJid = format!("bob@{domain}/web").parse().expect("jid");
+    let resumed_session = Session::new(&fresh_session.user_id, "bob", "bob");
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.to_string(),
+            user_id: format!("bob@{domain}"),
+            jid: detached_jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store");
+    state
+        .deps
+        .protocol
+        .resumable_sessions
+        .insert(stream_id.to_string(), resumed_session.clone());
+
+    let resume_frame = resume_frame_xml(stream_id, 0);
+    let responses = handle_xmpp_frame(&resume_frame, &domain, state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    let resumed = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(resumed.name(), "resumed");
+    assert!(matches!(
+        &conn.phase,
+        ConnectionPhase::Ready {
+            full_jid,
+            resumed: true,
+            ..
+        } if full_jid == &detached_jid
+    ));
+    assert_eq!(
+        conn.authenticated_session
+            .as_ref()
+            .map(|saved| saved.id.as_str()),
+        Some(resumed_session.id.as_str())
+    );
+    assert_ne!(
+        conn.authenticated_session
+            .as_ref()
+            .map(|saved| saved.id.as_str()),
+        Some(fresh_session.id.as_str())
+    );
+}
+
+#[tokio::test]
+async fn sm_resume_rejects_ready_phase() {
+    let state = create_test_websocket_state().await;
+    let mut conn = WsConnState::new();
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    conn.phase = ConnectionPhase::ready(jid, false);
+
+    let resume_frame = resume_frame_xml("stream-xyz", 0);
+    let responses =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "failed");
+    assert!(el
+        .get_child("unexpected-request", "urn:ietf:params:xml:ns:xmpp-stanzas")
+        .is_some());
+    assert!(matches!(conn.phase, ConnectionPhase::Ready { .. }));
+}
+
+#[tokio::test]
+async fn sm_enable_after_bind_returns_enabled_and_tracks_counters() {
+    let state = create_test_websocket_state().await;
+    let mut conn = WsConnState::new();
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    conn.phase = ConnectionPhase::ready(jid, false);
+
+    let responses = handle_xmpp_frame(
+        "<enable xmlns='urn:xmpp:sm:3' resume='true'/>",
+        "example.com",
+        state.as_ref(),
+        &mut conn,
+    )
+    .await;
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "enabled");
+    assert_eq!(el.attr("resume"), Some("true"));
+    assert!(el.attr("id").filter(|s| !s.is_empty()).is_some());
+    assert!(conn.sm_state.enabled);
+    assert!(conn.sm_state.is_resumable());
+
+    // An ack request bumps no counters but produces <a h=inbound_count/>.
+    let ack_responses = handle_xmpp_frame(
+        "<r xmlns='urn:xmpp:sm:3'/>",
+        "example.com",
+        state.as_ref(),
+        &mut conn,
+    )
+    .await;
+    assert_eq!(ack_responses.len(), 1);
+    let ack_el = Element::from_str(&ack_responses[0]).expect("xml");
+    assert_eq!(ack_el.name(), "a");
+    assert_eq!(ack_el.attr("h"), Some("0"));
+
+    // A countable inbound stanza bumps the inbound counter.
+    let _ = handle_xmpp_frame(
+        "<presence xmlns='jabber:client'/>",
+        "example.com",
+        state.as_ref(),
+        &mut conn,
+    )
+    .await;
+    assert_eq!(conn.sm_state.get_inbound_count(), 1);
+
+    // Subsequent <r/> should now report h=1.
+    let ack2 = handle_xmpp_frame(
+        "<r xmlns='urn:xmpp:sm:3'/>",
+        "example.com",
+        state.as_ref(),
+        &mut conn,
+    )
+    .await;
+    let ack2_el = Element::from_str(&ack2[0]).expect("xml");
+    assert_eq!(ack2_el.attr("h"), Some("1"));
+}
+
+#[tokio::test]
+async fn sm_resume_restores_session_and_replays_unacked() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    // Seed a detached session directly in the registry — this is the
+    // shape left behind by a prior WebSocket task after detach-on-close.
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    let stream_id = "stream-xyz".to_string();
+    let detached = DetachedSession {
+        stream_id: stream_id.clone(),
+        user_id: "alice@example.com".to_string(),
+        jid: jid.clone(),
+        inbound_count: 7,
+        outbound_count: 10,
+        last_acked: 8,
+        replay_gap_through: None,
+        unacked_stanzas: vec![
+            waddle_xmpp::stream_management::DetachedUnackedStanza {
+                sequence: 9,
+                stanza_xml: "<message id='m9'/>".to_string(),
+                original_receipt_at: chrono::Utc::now(),
+            },
+            waddle_xmpp::stream_management::DetachedUnackedStanza {
+                sequence: 10,
+                stanza_xml: "<message id='m10'/>".to_string(),
+                original_receipt_at: chrono::Utc::now(),
+            },
+        ],
+        max_resume_time: Some(300),
+        detached_at: std::time::Instant::now(),
+        carbons_enabled: true,
+        roster_interested: false,
+        presence_available: false,
+        presence_show: None,
+        presence_status: None,
+        presence_priority: 0,
+    };
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(detached)
+        .await
+        .expect("store");
+
+    let mut conn = WsConnState::new();
+    conn.phase = ConnectionPhase::authenticated(&jid);
+    // Client reports it has acked through 9, so only m10 needs replay.
+    let frame = resume_frame_xml(&stream_id, 9);
+    let responses = handle_xmpp_frame(&frame, "example.com", state.as_ref(), &mut conn).await;
+
+    // Expect <resumed/> first, then exactly the one unacked stanza.
+    assert!(!responses.is_empty());
+    let resumed = Element::from_str(&responses[0]).expect("resumed xml");
+    assert_eq!(resumed.name(), "resumed");
+    assert_eq!(resumed.attr("previd"), Some(stream_id.as_str()));
+
+    let replay_count = responses.len() - 1;
+    assert_eq!(
+        replay_count, 1,
+        "only m10 should be replayed: {responses:?}"
+    );
+    assert!(responses[1].contains("m10"));
+
+    // Session identity restored without SASL or bind frames.
+    assert!(conn.phase.is_authenticated());
+    assert!(conn.phase.is_ready());
+    assert_eq!(conn.phase.bound_jid(), Some(&jid));
+    assert!(conn.phase.is_resumed());
+    assert!(conn.carbons_enabled);
+    assert!(matches!(
+        &conn.phase,
+        ConnectionPhase::Ready {
+            full_jid,
+            resumed: true,
+            ..
+        } if full_jid == &jid
+    ));
+}
+
+#[tokio::test]
+async fn sm_resume_rejects_impossible_client_handled_count() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: "stream-too-far".to_string(),
+            user_id: "alice@example.com".to_string(),
+            jid: jid.clone(),
+            inbound_count: 4,
+            outbound_count: 2,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: vec![waddle_xmpp::stream_management::DetachedUnackedStanza {
+                sequence: 1,
+                stanza_xml: "<message id='m1'/>".to_string(),
+                original_receipt_at: chrono::Utc::now(),
+            }],
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store");
+
+    let mut conn = WsConnState::new();
+    conn.phase = ConnectionPhase::authenticated(&jid);
+    let resume_frame = resume_frame_xml("stream-too-far", 3);
+    let responses =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut conn).await;
+
+    assert_eq!(responses.len(), 1);
+    assert!(
+        responses[0].contains("stream:error")
+            && responses[0].contains("undefined-condition")
+            && responses[0].contains("handled-count-too-high")
+            && (responses[0].contains("h=\"3\"") || responses[0].contains("h='3'"))
+            && (responses[0].contains("send-count=\"2\"")
+                || responses[0].contains("send-count='2'")),
+        "invalid resume count should be a handled-count-too-high stream error: {responses:?}"
+    );
+    assert!(
+        !conn.sm_state.enabled,
+        "rejected resume must not pollute the fresh stream SM state"
+    );
+    assert!(
+        !conn.sm_state.is_resumable(),
+        "rejected resume must not make the fresh stream resumable"
+    );
+    assert!(
+        state
+            .deps
+            .protocol
+            .sm_session_registry
+            .take_session("stream-too-far")
+            .await
+            .expect("lookup")
+            .is_some(),
+        "rejected resume must release the detached session for a valid retry"
+    );
+}
+
+#[tokio::test]
+async fn sm_resume_replays_roster_push_recorded_while_detached() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    let stream_id = "stream-roster-replay".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: true,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store");
+
+    let recorded = state
+            .deps
+            .protocol
+            .sm_session_registry
+            .record_stanza_for_detached_resource(
+                &jid,
+                &Stanza::Iq(
+                    Element::from_str(
+                        "<iq xmlns='jabber:client' type='set' id='detached-roster-push'><query xmlns='jabber:iq:roster'/></iq>",
+                    )
+                    .expect("iq element")
+                    .try_into()
+                    .expect("iq stanza"),
+                ),
+                chrono::Utc::now(),
+            )
+            .await
+            .expect("record detached roster push");
+    assert!(recorded);
+
+    let mut conn = WsConnState::new();
+    conn.phase = ConnectionPhase::authenticated(&jid);
+    let frame = resume_frame_xml(&stream_id, 0);
+    let responses = handle_xmpp_frame(&frame, "example.com", state.as_ref(), &mut conn).await;
+
+    assert_eq!(
+        responses.len(),
+        2,
+        "expected resumed plus replay: {responses:?}"
+    );
+    assert!(responses[0].contains("<resumed"));
+    assert!(
+        responses[1].contains("detached-roster-push"),
+        "detached roster push should replay after resume: {responses:?}"
+    );
+    assert!(conn.roster_interested);
+}
+
+#[tokio::test]
+async fn direct_full_jid_message_records_for_detached_resource_replay() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_jid: FullJid = "alice@example.com/phone".parse().expect("alice jid");
+    let stream_id = "stream-detached-direct-message".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: alice_jid,
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached alice");
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    bob.ensure_state_machine(
+        "example.com",
+        &state.deps.protocol.dispatcher,
+        bob_jid,
+        false,
+        Blocklist::empty(),
+    );
+    let responses = handle_xmpp_frame(
+            r#"<message xmlns="jabber:client" type="chat" to="alice@example.com/phone" id="detached-dm-1"><body>queued while detached</body></message>"#,
+            "example.com",
+            state.as_ref(),
+            &mut bob,
+        )
+        .await;
+    assert!(responses.is_empty());
+
+    let detached = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session(&stream_id)
+        .await
+        .expect("take detached")
+        .expect("detached session remains");
+    assert!(
+        detached
+            .unacked_stanzas
+            .iter()
+            .any(|entry| entry.stanza_xml.contains("detached-dm-1")),
+        "full-JID direct message should be recorded for detached replay: {detached:?}"
+    );
+}
+
+#[tokio::test]
+async fn bare_jid_message_records_for_detached_resource_replay() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_jid: FullJid = "alice@example.com/phone".parse().expect("alice jid");
+    let stream_id = "stream-detached-bare-message".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: alice_jid,
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached alice");
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    bob.ensure_state_machine(
+        "example.com",
+        &state.deps.protocol.dispatcher,
+        bob_jid,
+        false,
+        Blocklist::empty(),
+    );
+    let responses = handle_xmpp_frame(
+            r#"<message xmlns="jabber:client" type="chat" to="alice@example.com" id="detached-bare-dm-1"><body>queued while detached</body></message>"#,
+            "example.com",
+            state.as_ref(),
+            &mut bob,
+        )
+        .await;
+    assert!(responses.is_empty());
+
+    let detached = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session(&stream_id)
+        .await
+        .expect("take detached")
+        .expect("detached session remains");
+    assert!(
+        detached
+            .unacked_stanzas
+            .iter()
+            .any(|entry| entry.stanza_xml.contains("detached-bare-dm-1")),
+        "bare-JID direct message should be recorded for detached replay: {detached:?}"
+    );
+    // RFC 6121 §8.5.2.1.1: bare-JID delivery routes the original
+    // stanza to each available resource without rewriting `to`.
+    // The dispatcher path preserves this — legacy `handle_message`
+    // rewrote `to` to the per-resource full JID, which was a
+    // server-side deviation from the RFC. Assert only the
+    // reachability semantic here; integration tests verify the
+    // wire shape end-to-end.
+}
+
+#[tokio::test]
+async fn message_carbons_record_for_detached_enabled_resources() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let alice_phone: FullJid = "alice@example.com/phone".parse().expect("alice phone");
+    let alice_laptop: FullJid = "alice@example.com/laptop".parse().expect("alice laptop");
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let sent_stream_id = "stream-detached-sent-carbon".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: sent_stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: alice_laptop.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: true,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached alice laptop");
+
+    let mut alice = WsConnState::new();
+    alice.phase = ConnectionPhase::ready(alice_phone.clone(), false);
+    alice.ensure_state_machine(
+        "example.com",
+        &state.deps.protocol.dispatcher,
+        alice_phone.clone(),
+        false,
+        Blocklist::empty(),
+    );
+    let responses = handle_xmpp_frame(
+            r#"<message xmlns="jabber:client" type="chat" to="bob@example.com/web" id="detached-sent-carbon-source"><body>copy me</body></message>"#,
+            "example.com",
+            state.as_ref(),
+            &mut alice,
+        )
+        .await;
+    assert!(responses.is_empty());
+
+    let sent_detached = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session(&sent_stream_id)
+        .await
+        .expect("take sent detached")
+        .expect("sent detached session remains");
+    assert!(
+        sent_detached
+            .unacked_stanzas
+            .iter()
+            .any(|entry| entry.stanza_xml.contains("<sent")
+                && entry.stanza_xml.contains("urn:xmpp:carbons:2")
+                && entry.stanza_xml.contains("detached-sent-carbon-source")),
+        "sent carbon should be recorded for detached opted-in resource: {sent_detached:?}"
+    );
+
+    let received_stream_id = "stream-detached-received-carbon".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: received_stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: alice_laptop,
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: true,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached alice laptop again");
+
+    // Build alice/phone's per-connection state machine so we can
+    // drive the recipient-pass carbon fan-out the dispatcher path
+    // owns. In production this happens automatically via
+    // alice/phone's main loop dispatching the queued
+    // `DeliveryKind::PeerStanza`; the unit test reproduces the
+    // same step explicitly.
+    let mut alice_phone_conn = WsConnState::new();
+    alice_phone_conn.phase = ConnectionPhase::ready(alice_phone.clone(), false);
+    alice_phone_conn.ensure_state_machine(
+        "example.com",
+        &state.deps.protocol.dispatcher,
+        alice_phone.clone(),
+        false,
+        Blocklist::empty(),
+    );
+    let (alice_phone_tx, mut alice_phone_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(alice_phone.clone(), alice_phone_tx);
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    bob.ensure_state_machine(
+        "example.com",
+        &state.deps.protocol.dispatcher,
+        bob_jid,
+        false,
+        Blocklist::empty(),
+    );
+    let responses = handle_xmpp_frame(
+            r#"<message xmlns="jabber:client" type="chat" to="alice@example.com/phone" id="detached-received-carbon-source"><body>copy me too</body></message>"#,
+            "example.com",
+            state.as_ref(),
+            &mut bob,
+        )
+        .await;
+    assert!(responses.is_empty());
+
+    // Pump the queued PeerStanza through alice/phone's SM so the
+    // recipient pass runs and the dispatcher emits the
+    // received-carbon fan-out. This is the same dispatch the
+    // production main loop performs on `DeliveryKind::PeerStanza`.
+    while let Ok(outbound) = alice_phone_rx.try_recv() {
+        if !matches!(outbound.kind, DeliveryKind::PeerStanza) {
+            continue;
+        }
+        let sm = alice_phone_conn.state_machine.as_mut().expect("alice SM");
+        let events = sm.handle(InboundEvent::StanzaFromPeer(Box::new(outbound.stanza)));
+        let deps = build_interpret_deps(state.as_ref(), None);
+        let _ = drive_interpret_loop(events, sm, &deps).await;
+    }
+
+    let received_detached = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session(&received_stream_id)
+        .await
+        .expect("take received detached")
+        .expect("received detached session remains");
+    assert!(
+        received_detached.unacked_stanzas.iter().any(|entry| entry
+            .stanza_xml
+            .contains("<received")
+            && entry.stanza_xml.contains("urn:xmpp:carbons:2")
+            && entry.stanza_xml.contains("detached-received-carbon-source")),
+        "received carbon should be recorded for detached opted-in resource: {received_detached:?}"
+    );
+}
+
+#[tokio::test]
+async fn duplicate_subscribe_ack_reaches_non_roster_interested_resource() {
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_jid: FullJid = "alice@example.com/phone".parse().expect("alice jid");
+    let (bob_tx, mut bob_rx) = mpsc::channel::<OutboundStanza>(16);
+    let (alice_tx, mut alice_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(bob_jid.clone(), bob_tx);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(alice_jid.clone(), alice_tx);
+
+    let mut alice = WsConnState::new();
+    alice.phase = ConnectionPhase::ready(alice_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+            r#"<iq xmlns="jabber:client" type="get" id="alice-roster"><query xmlns="jabber:iq:roster"/></iq>"#,
+            "example.com",
+            state.as_ref(),
+            &mut alice,
+        )
+        .await;
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice,
+    )
+    .await;
+    while alice_rx.try_recv().is_ok() {}
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    let _ = tokio::time::timeout(std::time::Duration::from_millis(250), alice_rx.recv())
+        .await
+        .expect("alice receives initial subscribe")
+        .expect("subscribe stanza");
+
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribed" to="bob@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice,
+    )
+    .await;
+    while bob_rx.try_recv().is_ok() {}
+
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    let ack = tokio::time::timeout(std::time::Duration::from_millis(250), bob_rx.recv())
+        .await
+        .expect("duplicate subscribe ack")
+        .expect("ack stanza");
+    let frame = stanza_to_xml(&ack.stanza);
+    assert!(
+        frame.contains("from=\"alice@example.com\"")
+            && frame.contains("to=\"bob@example.com\"")
+            && frame.contains("type=\"subscribed\""),
+        "duplicate subscribe ack should reach a live resource even before roster get: {frame}"
+    );
+}
+
+#[tokio::test]
+async fn roster_set_records_push_for_detached_interested_resource() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let detached_jid: FullJid = "alice@example.com/web".parse().expect("detached jid");
+    let source_jid: FullJid = "alice@example.com/phone".parse().expect("source jid");
+    let stream_id = "stream-roster-fanout".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: detached_jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: true,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached session");
+
+    let mut source = WsConnState::new();
+    source.phase = ConnectionPhase::ready(source_jid, false);
+    let responses = handle_xmpp_frame(
+            r#"<iq xmlns="jabber:client" type="set" id="roster-detached-fanout"><query xmlns="jabber:iq:roster"><item jid="bob@example.com" name="Bob"/></query></iq>"#,
+            "example.com",
+            state.as_ref(),
+            &mut source,
+        )
+        .await;
+    assert!(
+        responses
+            .iter()
+            .any(|frame| frame.contains("roster-detached-fanout")
+                && frame.contains("type=\"result\"")),
+        "roster set should succeed: {responses:?}"
+    );
+
+    let mut resumed = WsConnState::new();
+    resumed.phase = ConnectionPhase::authenticated(&detached_jid);
+    let resume_frame = resume_frame_xml(&stream_id, 0);
+    let replay =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut resumed).await;
+    assert!(
+        replay
+            .iter()
+            .any(|frame| frame.contains("jabber:iq:roster") && frame.contains("bob@example.com")),
+        "detached interested resource should replay roster fanout push: {replay:?}"
+    );
+    assert!(resumed.roster_interested);
+}
+
+#[tokio::test]
+async fn subscription_approval_replays_current_presence_from_detached_available_resource() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_web_jid: FullJid = "alice@example.com/web".parse().expect("alice web jid");
+    let alice_phone_jid: FullJid = "alice@example.com/phone".parse().expect("alice phone jid");
+    let (bob_tx, mut bob_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(bob_jid.clone(), bob_tx);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .mark_roster_interested(&bob_jid);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .update_presence(&bob_jid, true, 0);
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    while bob_rx.try_recv().is_ok() {}
+
+    let stream_id = "stream-detached-current-presence".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id,
+            user_id: "alice@example.com".to_string(),
+            jid: alice_web_jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: true,
+            presence_show: Some(xmpp_parsers::presence::Show::Chat),
+            presence_status: Some("ready from detach".to_string()),
+            presence_priority: 7,
+        })
+        .await
+        .expect("store detached alice web");
+
+    let mut alice_phone = WsConnState::new();
+    alice_phone.phase = ConnectionPhase::ready(alice_phone_jid, false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribed" to="bob@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice_phone,
+    )
+    .await;
+
+    let mut delivered = Vec::new();
+    for _ in 0..4 {
+        if let Ok(Some(outbound)) =
+            tokio::time::timeout(std::time::Duration::from_millis(250), bob_rx.recv()).await
+        {
+            delivered.push(stanza_to_xml(&outbound.stanza));
+        }
+    }
+    assert!(
+        delivered.iter().any(|frame| {
+            frame.contains("from=\"alice@example.com/web\"")
+                && frame.contains("<show>chat</show>")
+                && frame.contains("<status>ready from detach</status>")
+                && frame.contains("<priority>7</priority>")
+        }),
+        "approval should deliver current rich presence from detached available resource: {delivered:?}"
+    );
+}
+
+#[tokio::test]
+async fn presence_probe_returns_detached_available_resource_presence() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_jid: FullJid = "alice@example.com/phone".parse().expect("alice jid");
+    let (bob_tx, mut bob_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(bob_jid.clone(), bob_tx);
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    let mut alice = WsConnState::new();
+    alice.phase = ConnectionPhase::ready(alice_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribed" to="bob@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice,
+    )
+    .await;
+    while bob_rx.try_recv().is_ok() {}
+
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: "stream-detached-probe".to_string(),
+            user_id: "alice@example.com".to_string(),
+            jid: alice_jid,
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: true,
+            presence_show: Some(xmpp_parsers::presence::Show::Away),
+            presence_status: Some("stepped away".to_string()),
+            presence_priority: 5,
+        })
+        .await
+        .expect("store detached alice");
+
+    bob.phase = ConnectionPhase::ready(bob_jid, false);
+    let responses = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="probe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    assert!(responses.is_empty());
+
+    let outbound = tokio::time::timeout(std::time::Duration::from_millis(250), bob_rx.recv())
+        .await
+        .expect("probe response")
+        .expect("outbound stanza");
+    let frame = stanza_to_xml(&outbound.stanza);
+    assert!(
+        frame.contains("from=\"alice@example.com/phone\"")
+            && frame.contains("to=\"bob@example.com\"")
+            && frame.contains("<show>away</show>")
+            && frame.contains("<status>stepped away</status>")
+            && frame.contains("<priority>5</priority>"),
+        "probe should return rich presence from detached available resource: {frame}"
+    );
+}
+
+#[tokio::test]
+async fn full_jid_presence_probe_returns_only_that_resources_availability() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_phone: FullJid = "alice@example.com/phone".parse().expect("alice phone");
+    let alice_tablet: FullJid = "alice@example.com/tablet".parse().expect("alice tablet");
+    let (bob_tx, mut bob_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(bob_jid.clone(), bob_tx);
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    let mut alice = WsConnState::new();
+    alice.phase = ConnectionPhase::ready(alice_phone.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribed" to="bob@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice,
+    )
+    .await;
+    while bob_rx.try_recv().is_ok() {}
+
+    for (stream_id, jid, show, status) in [
+        (
+            "stream-probe-phone",
+            alice_phone.clone(),
+            xmpp_parsers::presence::Show::Away,
+            "phone detail",
+        ),
+        (
+            "stream-probe-tablet",
+            alice_tablet,
+            xmpp_parsers::presence::Show::Chat,
+            "tablet detail",
+        ),
+    ] {
+        state
+            .deps
+            .protocol
+            .sm_session_registry
+            .store_session(DetachedSession {
+                stream_id: stream_id.to_string(),
+                user_id: "alice@example.com".to_string(),
+                jid,
+                inbound_count: 0,
+                outbound_count: 0,
+                last_acked: 0,
+                replay_gap_through: None,
+                unacked_stanzas: Vec::new(),
+                max_resume_time: Some(300),
+                detached_at: std::time::Instant::now(),
+                carbons_enabled: false,
+                roster_interested: false,
+                presence_available: true,
+                presence_show: Some(show),
+                presence_status: Some(status.to_string()),
+                presence_priority: 5,
+            })
+            .await
+            .expect("store detached alice resource");
+    }
+
+    let responses = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="probe" to="alice@example.com/phone"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    assert!(responses.is_empty());
+
+    let outbound = tokio::time::timeout(std::time::Duration::from_millis(250), bob_rx.recv())
+        .await
+        .expect("full-jid probe response")
+        .expect("outbound stanza");
+    let frame = stanza_to_xml(&outbound.stanza);
+    assert!(
+        frame.contains("from=\"alice@example.com/phone\"")
+            && frame.contains("to=\"bob@example.com\"")
+            && frame.contains("<show>away</show>")
+            && frame.contains("<status>phone detail</status>")
+            && frame.contains("<priority>5</priority>")
+            && !frame.contains("alice@example.com/tablet"),
+        "full-JID probe should return rich presence only for the requested resource: {frame}"
+    );
+    assert!(
+        tokio::time::timeout(std::time::Duration::from_millis(100), bob_rx.recv())
+            .await
+            .is_err(),
+        "full-JID probe must not return sibling resources"
+    );
+}
+
+#[tokio::test]
+async fn presence_probe_without_subscription_does_not_reveal_detached_presence() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let mallory_jid: FullJid = "mallory@example.com/web".parse().expect("mallory jid");
+    let alice_jid: FullJid = "alice@example.com/phone".parse().expect("alice jid");
+    let (mallory_tx, mut mallory_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(mallory_jid.clone(), mallory_tx);
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: "stream-detached-probe-denied".to_string(),
+            user_id: "alice@example.com".to_string(),
+            jid: alice_jid,
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: true,
+            presence_show: Some(xmpp_parsers::presence::Show::Away),
+            presence_status: Some("private".to_string()),
+            presence_priority: 5,
+        })
+        .await
+        .expect("store detached alice");
+
+    let mut mallory = WsConnState::new();
+    mallory.phase = ConnectionPhase::ready(mallory_jid, false);
+    let responses = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="probe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut mallory,
+    )
+    .await;
+    assert!(responses.is_empty());
+    let outbound = tokio::time::timeout(std::time::Duration::from_millis(250), mallory_rx.recv())
+        .await
+        .expect("unsubscribed probe response")
+        .expect("outbound stanza");
+    let frame = stanza_to_xml(&outbound.stanza);
+    assert!(
+        frame.contains("from=\"alice@example.com\"")
+            && frame.contains("to=\"mallory@example.com\"")
+            && frame.contains("type=\"unsubscribed\"")
+            && !frame.contains("alice@example.com/phone")
+            && !frame.contains("private"),
+        "unauthorized probe must return only an unsubscribed signal: {frame}"
+    );
+}
+
+#[tokio::test]
+async fn expired_detached_available_session_broadcasts_unavailable_to_subscribers() {
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_jid: FullJid = "alice@example.com/phone".parse().expect("alice jid");
+    let alice_sibling_jid: FullJid = "alice@example.com/laptop".parse().expect("alice sibling");
+    let (bob_tx, mut bob_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(bob_jid.clone(), bob_tx);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .update_presence(&bob_jid, true, 0);
+    let (alice_sibling_tx, mut alice_sibling_rx) = mpsc::channel::<OutboundStanza>(16);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .register(alice_sibling_jid.clone(), alice_sibling_tx);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .update_presence(&alice_sibling_jid, true, 0);
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid, false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    let mut alice = WsConnState::new();
+    alice.phase = ConnectionPhase::ready(alice_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribed" to="bob@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice,
+    )
+    .await;
+    while bob_rx.try_recv().is_ok() {}
+    while alice_sibling_rx.try_recv().is_ok() {}
+
+    handlers::presence::broadcast_unavailable_for_expired_detached_session(
+        state.as_ref(),
+        &alice_jid,
+    )
+    .await;
+
+    let outbound = tokio::time::timeout(std::time::Duration::from_millis(250), bob_rx.recv())
+        .await
+        .expect("unavailable broadcast")
+        .expect("outbound stanza");
+    let frame = stanza_to_xml(&outbound.stanza);
+    assert!(
+        frame.contains("from=\"alice@example.com/phone\"")
+            && frame.contains("to=\"bob@example.com\"")
+            && frame.contains("type=\"unavailable\""),
+        "expired detached session should broadcast unavailable presence: {frame}"
+    );
+    let sibling_outbound = tokio::time::timeout(
+        std::time::Duration::from_millis(250),
+        alice_sibling_rx.recv(),
+    )
+    .await
+    .expect("sibling unavailable broadcast")
+    .expect("outbound stanza");
+    let sibling_frame = stanza_to_xml(&sibling_outbound.stanza);
+    assert!(
+        sibling_frame.contains("from=\"alice@example.com/phone\"")
+            && sibling_frame.contains("to=\"alice@example.com\"")
+            && sibling_frame.contains("type=\"unavailable\""),
+        "expired detached session should notify sibling resources: {sibling_frame}"
+    );
+}
+
+#[tokio::test]
+async fn subscription_approval_records_roster_push_for_detached_interested_resource() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_jid: FullJid = "alice@example.com/web".parse().expect("alice jid");
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+            r#"<iq xmlns="jabber:client" type="get" id="bob-roster"><query xmlns="jabber:iq:roster"/></iq>"#,
+            "example.com",
+            state.as_ref(),
+            &mut bob,
+        )
+        .await;
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+
+    let stream_id = "stream-detached-subscription-roster-push".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "bob@example.com".to_string(),
+            jid: bob_jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: true,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached bob");
+
+    let mut alice = WsConnState::new();
+    alice.phase = ConnectionPhase::ready(alice_jid, false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribed" to="bob@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice,
+    )
+    .await;
+
+    let mut resumed = WsConnState::new();
+    resumed.phase = ConnectionPhase::authenticated(&bob_jid);
+    let resume_frame = resume_frame_xml(&stream_id, 0);
+    let replay =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut resumed).await;
+    assert!(
+        replay.iter().any(|frame| {
+            frame.contains("jabber:iq:roster")
+                && frame.contains("alice@example.com")
+                && frame.contains("subscription=\"to\"")
+        }),
+        "detached interested resource should replay subscription roster push: {replay:?}"
+    );
+}
+
+#[tokio::test]
+async fn subscribe_to_detached_available_resource_replays_on_resume() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let alice_jid: FullJid = "alice@example.com/web".parse().expect("alice jid");
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let stream_id = "stream-detached-subscribe-recipient".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: alice_jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: true,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached alice");
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid, false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+
+    let mut resumed = WsConnState::new();
+    resumed.phase = ConnectionPhase::authenticated(&alice_jid);
+    let resume_frame = resume_frame_xml(&stream_id, 0);
+    let replay =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut resumed).await;
+    assert!(
+        replay.iter().any(|frame| {
+            frame.contains("type=\"subscribe\"") && frame.contains("from=\"bob@example.com\"")
+        }),
+        "detached available recipient should replay inbound subscribe: {replay:?}"
+    );
+}
+
+#[tokio::test]
+async fn presence_broadcast_to_detached_available_subscriber_replays_on_resume() {
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let bob_jid: FullJid = "bob@example.com/web".parse().expect("bob jid");
+    let alice_jid: FullJid = "alice@example.com/web".parse().expect("alice jid");
+
+    let mut bob = WsConnState::new();
+    bob.phase = ConnectionPhase::ready(bob_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribe" to="alice@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut bob,
+    )
+    .await;
+    let mut alice = WsConnState::new();
+    alice.phase = ConnectionPhase::ready(alice_jid.clone(), false);
+    let _ = handle_xmpp_frame(
+        r#"<presence xmlns="jabber:client" type="subscribed" to="bob@example.com"/>"#,
+        "example.com",
+        state.as_ref(),
+        &mut alice,
+    )
+    .await;
+
+    let stream_id = "stream-detached-presence-broadcast".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "bob@example.com".to_string(),
+            jid: bob_jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(300),
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: true,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store detached bob");
+
+    let _ = handle_xmpp_frame(
+            r#"<presence xmlns="jabber:client"><show>away</show><status>broadcast while detached</status><priority>5</priority></presence>"#,
+            "example.com",
+            state.as_ref(),
+            &mut alice,
+        )
+        .await;
+
+    let mut resumed = WsConnState::new();
+    resumed.phase = ConnectionPhase::authenticated(&bob_jid);
+    let resume_frame = resume_frame_xml(&stream_id, 0);
+    let replay =
+        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut resumed).await;
+    assert!(
+        replay.iter().any(|frame| {
+            frame.contains("from=\"alice@example.com/web\"")
+                && frame.contains("<show>away</show>")
+                && frame.contains("<status>broadcast while detached</status>")
+                && frame.contains("<priority>5</priority>")
+        }),
+        "detached available subscriber should replay presence broadcast: {replay:?}"
+    );
+}
+
+#[tokio::test]
+async fn sm_resume_with_unknown_stream_id_fails() {
+    let state = create_test_websocket_state().await;
+    let mut conn = WsConnState::new();
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    conn.phase = ConnectionPhase::authenticated(&jid);
+    let frame = resume_frame_xml("does-not-exist", 0);
+    let responses = handle_xmpp_frame(&frame, "example.com", state.as_ref(), &mut conn).await;
+    assert_eq!(responses.len(), 1);
+    let el = Element::from_str(&responses[0]).expect("xml");
+    assert_eq!(el.name(), "failed");
+    // Must NOT mark the session as bound/resumed.
+    assert!(conn.phase.is_authenticated());
+    assert!(!conn.phase.is_ready());
+    assert!(!conn.phase.is_resumed());
+}
+
+#[tokio::test]
+async fn sm_resume_signals_suppress_record_so_main_loop_skips_replay() {
+    // Regression guard for the double-record bug reported in PR review:
+    // `handle_sm_resume` must request suppression of outbound recording
+    // for its own response batch. Replayed stanzas are already in the
+    // unacked queue — re-recording them would bump `outbound_count` and
+    // create duplicates.
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    let stream_id = "stream-dup-check".to_string();
+    let detached = DetachedSession {
+        stream_id: stream_id.clone(),
+        user_id: "alice@example.com".to_string(),
+        jid: jid.clone(),
+        inbound_count: 0,
+        outbound_count: 2,
+        last_acked: 0,
+        replay_gap_through: None,
+        unacked_stanzas: vec![
+            waddle_xmpp::stream_management::DetachedUnackedStanza {
+                sequence: 1,
+                stanza_xml: "<message id='m1'/>".to_string(),
+                original_receipt_at: chrono::Utc::now(),
+            },
+            waddle_xmpp::stream_management::DetachedUnackedStanza {
+                sequence: 2,
+                stanza_xml: "<message id='m2'/>".to_string(),
+                original_receipt_at: chrono::Utc::now(),
+            },
+        ],
+        max_resume_time: Some(300),
+        detached_at: std::time::Instant::now(),
+        carbons_enabled: false,
+        roster_interested: false,
+        presence_available: false,
+        presence_show: None,
+        presence_status: None,
+        presence_priority: 0,
+    };
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(detached)
+        .await
+        .expect("store");
+
+    let mut conn = WsConnState::new();
+    conn.phase = ConnectionPhase::authenticated(&jid);
+    let frame = resume_frame_xml(&stream_id, 0);
+    let _ = handle_xmpp_frame(&frame, "example.com", state.as_ref(), &mut conn).await;
+
+    // The resume handler must have raised the suppress flag so the main
+    // loop skips re-recording its own response batch.
+    assert!(
+        conn.suppress_sm_record_next_batch,
+        "handle_sm_resume must ask the main loop to skip SM recording for this batch"
+    );
+    // And the restored counters must still reflect what the client had
+    // acknowledged, not the inflated post-re-record values (2, not 4).
+    assert_eq!(conn.sm_state.outbound_count, 2);
+    assert_eq!(conn.sm_state.queue_len(), 2);
+}
+
+#[tokio::test]
+async fn cleanup_shutdown_detaches_resumable_session_on_transport_drop() {
+    let state = create_test_websocket_state().await;
+    let owner_session = create_test_server_owner_session(state.as_ref(), "alice").await;
+    let room_jid: BareJid = "detached-channel@muc.example.com".parse().expect("room");
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+
+    let _ = handle_muc_join(
+        state.as_ref(),
+        "example.com",
+        &room_jid,
+        &jid,
+        "alice",
+        &Some(owner_session),
+    )
+    .await;
+    let (tx, mut rx) = mpsc::channel::<OutboundStanza>(4);
+    let owner = state
+        .deps
+        .protocol
+        .connection_registry
+        .register(jid.clone(), tx);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .update_presence(&jid, true, 0);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .update_presence_state(
+            &jid,
+            Some("away".to_string()),
+            Some("stepped out".to_string()),
+            3,
+        );
+
+    let mut conn = WsConnState::new();
+    conn.phase = ConnectionPhase::ready(jid.clone(), false);
+    conn.registry_owner = Some(owner);
+    conn.roster_interested = true;
+    conn.sm_state
+        .enable("stream-detach".to_string(), true, Some(300));
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .send_to(
+            &jid,
+            Stanza::Presence(xmpp_parsers::presence::Presence::new(
+                xmpp_parsers::presence::Type::None,
+            )),
+        )
+        .await;
+
+    cleanup_connection_shutdown(state.as_ref(), &mut rx, &mut conn, false).await;
+
+    assert!(!state.deps.protocol.connection_registry.is_connected(&jid));
+    let detached = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session("stream-detach")
+        .await
+        .expect("registry lookup");
+    let detached = detached.expect("detached session");
+    assert!(
+        detached.roster_interested,
+        "detached session must preserve roster-interest state"
+    );
+    assert!(
+        detached.presence_available,
+        "detached session must preserve available-presence state"
+    );
+    assert_eq!(
+        detached.presence_show,
+        Some(xmpp_parsers::presence::Show::Away)
+    );
+    assert_eq!(detached.presence_status.as_deref(), Some("stepped out"));
+    assert_eq!(detached.presence_priority, 3);
+    assert!(
+        detached
+            .unacked_stanzas
+            .iter()
+            .any(|entry| entry.stanza_xml.contains("<presence")),
+        "cleanup must record queued-but-unwritten outbound stanzas before detaching"
+    );
+    assert!(snapshot_room(state.as_ref(), &room_jid)
+        .await
+        .room
+        .find_nick_by_real_jid(&jid)
+        .is_some());
+}
+
+#[tokio::test]
+async fn cleanup_shutdown_does_not_detach_explicit_close() {
+    let state = create_test_websocket_state().await;
+    let owner_session = create_test_server_owner_session(state.as_ref(), "alice").await;
+    let room_jid: BareJid = "closing-channel@muc.example.com".parse().expect("room");
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+
+    let _ = handle_muc_join(
+        state.as_ref(),
+        "example.com",
+        &room_jid,
+        &jid,
+        "alice",
+        &Some(owner_session),
+    )
+    .await;
+    let (tx, mut rx) = mpsc::channel::<OutboundStanza>(4);
+    let owner = state
+        .deps
+        .protocol
+        .connection_registry
+        .register(jid.clone(), tx);
+
+    let mut conn = WsConnState::new();
+    conn.phase = ConnectionPhase::ready(jid.clone(), false);
+    conn.registry_owner = Some(owner);
+    conn.sm_state
+        .enable("stream-close".to_string(), false, Some(300));
+
+    cleanup_connection_shutdown(state.as_ref(), &mut rx, &mut conn, false).await;
+
+    assert!(!state.deps.protocol.connection_registry.is_connected(&jid));
+    let detached = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .take_session("stream-close")
+        .await
+        .expect("registry lookup");
+    assert!(
+        detached.is_none(),
+        "explicit <close/> must not leave a resumable detached session behind"
+    );
+    assert!(snapshot_room(state.as_ref(), &room_jid)
+        .await
+        .room
+        .find_nick_by_real_jid(&jid)
+        .is_none());
+}
+
+#[tokio::test]
+async fn cleanup_shutdown_does_not_unregister_replacement_session() {
+    let state = create_test_websocket_state().await;
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+    let (old_tx, mut old_rx) = mpsc::channel::<OutboundStanza>(4);
+    let (new_tx, _new_rx) = mpsc::channel::<OutboundStanza>(4);
+
+    let old_owner = state
+        .deps
+        .protocol
+        .connection_registry
+        .register(jid.clone(), old_tx);
+    let new_owner = state
+        .deps
+        .protocol
+        .connection_registry
+        .register(jid.clone(), new_tx);
+
+    let mut old_conn = WsConnState::new();
+    old_conn.phase = ConnectionPhase::ready(jid.clone(), false);
+    old_conn.registry_owner = Some(old_owner);
+
+    cleanup_connection_shutdown(state.as_ref(), &mut old_rx, &mut old_conn, false).await;
+
+    assert!(
+        state.deps.protocol.connection_registry.is_connected(&jid),
+        "cleanup for a replaced connection must leave the replacement registered"
+    );
+    assert!(
+        state
+            .deps
+            .protocol
+            .connection_registry
+            .unregister_if_owner(&jid, &new_owner)
+            .is_some(),
+        "the remaining registry owner should be the replacement session"
+    );
+}
+
+#[tokio::test]
+async fn sm_janitor_helper_drains_expired_and_cleans_muc() {
+    // Exercise the pieces the janitor composes: drain_expired() returns
+    // the removed sessions, and cleanup_muc_presence_for_jid removes the
+    // occupant that was held while the session was detached.
+    use waddle_xmpp::stream_management::{DetachedSession, SmSessionRegistry};
+    let state = create_test_websocket_state().await;
+    let owner_session = create_test_server_owner_session(state.as_ref(), "alice").await;
+    let room_jid: BareJid = "expired-channel@muc.example.com".parse().expect("room");
+    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
+
+    // Put alice in the room, as if she'd detached with SM.
+    let _ = handle_muc_join(
+        state.as_ref(),
+        "example.com",
+        &room_jid,
+        &jid,
+        "alice",
+        &Some(owner_session),
+    )
+    .await;
+    assert!(snapshot_room(state.as_ref(), &room_jid)
+        .await
+        .room
+        .find_nick_by_real_jid(&jid)
+        .is_some());
+
+    // Seed an immediately-expired detached session for that JID.
+    let stream_id = "already-expired".to_string();
+    state
+        .deps
+        .protocol
+        .sm_session_registry
+        .store_session(DetachedSession {
+            stream_id: stream_id.clone(),
+            user_id: "alice@example.com".to_string(),
+            jid: jid.clone(),
+            inbound_count: 0,
+            outbound_count: 0,
+            last_acked: 0,
+            replay_gap_through: None,
+            unacked_stanzas: Vec::new(),
+            max_resume_time: Some(0), // already expired
+            detached_at: std::time::Instant::now(),
+            carbons_enabled: false,
+            roster_interested: false,
+            presence_available: false,
+            presence_show: None,
+            presence_status: None,
+            presence_priority: 0,
+        })
+        .await
+        .expect("store");
+    state
+        .deps
+        .protocol
+        .resumable_sessions
+        .insert(stream_id.clone(), Session::new("uid", "alice", "alice"));
+
+    // Wait a hair so the 0-second TTL is definitely in the past.
+    tokio::time::sleep(std::time::Duration::from_millis(10)).await;
+
+    let drained = state
+        .deps
+        .protocol
+        .sm_session_registry
+        .drain_expired()
+        .await
+        .expect("drain");
+    assert_eq!(drained.len(), 1);
+    assert_eq!(drained[0].stream_id, stream_id);
+
+    // The janitor body: remove sidecar + MUC occupant + any routing slot.
+    state.deps.protocol.resumable_sessions.remove(&stream_id);
+    state
+        .deps
+        .protocol
+        .connection_registry
+        .unregister(&drained[0].jid);
+    cleanup_muc_presence_for_jid(state.as_ref(), &drained[0].jid).await;
+
+    assert!(!state
+        .deps
+        .protocol
+        .resumable_sessions
+        .contains_key(&stream_id));
+    assert!(
+        snapshot_room(state.as_ref(), &room_jid)
+            .await
+            .room
+            .find_nick_by_real_jid(&jid)
+            .is_none(),
+        "MUC occupant must be gone after janitor sweep"
+    );
+}
